This is pretty important commit to fix #41 bug.
It is pretty important to filter out multi-mapping reads before
doing reads against features counts before differential expression
analysis. As you can imagine including multi-mapping reads exaggerates
gene expression.
It is a little bit hard to know whether this will be a huge problem
for differential expression analysis. Perhaps if multi-mappers are
consistent between samples and because we are measuring relative
difference that isn't a big problem. Also bear in mind if a read
multi-maps to a location that doesn't have a know feature then that
read will discarded and no information recorded in the table of counts.

Refer to actual issue #41 for more information about MAPQ field and how
different aligners treat it.